Tests/LibWeb/Text/expected/HTML/link-element-rapid-href-parser-created.txt

Starting parser-created element test...
Survived parser-created element stress test
